fix: change to u32

+++ async-claude/src/messages/request.rs
@@ -8,7 +8,7 @@ pub struct Request {
     pub messages: Vec<Message>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub system: Option<String>,
-    pub max_tokens: usize,
+    pub max_tokens: u32,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub metadata: Option<serde_json::Value>,
     #[serde(skip_serializing_if = "Option::is_none")]
